Dangle iterators code on Visual Studio
According to cppreference, push_back has the following behaviour against
iterators:
If the new size() is greater than capacity() then all iterators and
references (including the past-the-end iterator) are invalidated.
Otherwise only the past-the-end iterator is invalidated.

Thus std::rend() and std::end() might be invalidated after a push_back.
It seems to be the case with VS but not with G++ and Clang.